requirements-test.txt
djangorestframework
codecov
coverage
pre-commit
tox
